requirements.txt
future==0.16.0
beautifulsoup4==4.6.0
Django==2.0.2
django-bootstrap4==0.0.6
django-compressor==2.2
django-extensions==2.0.0
django-taggit==0.22.2
future==0.16.0
psycopg2==2.7.4
requests==2.18.4